Time: 4 ms (98.93%), Space: 5.9 MB (44.13%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
             return false;
         }
         unsigned int result = 0;
-        unsigned int remainder;
+        short remainder;
         unsigned int y = x;
         while(y != 0){
             remainder = y % 10;
